+++ tests/vipers-tests/src/lib.rs
@@ -1,25 +1,5 @@
-//! Vipers tests for Anchor 0.20.1.
+//! Vipers tests.
 #![cfg(test)]
-
-#[cfg(feature = "anchor-0_21_0")]
-extern crate anchor_lang_0_21_0 as anchor_lang;
-#[cfg(feature = "anchor-0_21_0")]
-extern crate anchor_spl_0_21_0 as anchor_spl;
-
-#[cfg(feature = "anchor-0_20_1")]
-extern crate anchor_lang_0_20_1 as anchor_lang;
-#[cfg(feature = "anchor-0_20_1")]
-extern crate anchor_spl_0_20_1 as anchor_spl;
-
-#[cfg(feature = "anchor-0_19_0")]
-extern crate anchor_lang_0_19_0 as anchor_lang;
-#[cfg(feature = "anchor-0_19_0")]
-extern crate anchor_spl_0_19_0 as anchor_spl;
-
-#[cfg(feature = "anchor-0_18_2")]
-extern crate anchor_lang_0_18_2 as anchor_lang;
-#[cfg(feature = "anchor-0_18_2")]
-extern crate anchor_spl_0_18_2 as anchor_spl;
 
 use anchor_lang::prelude::*;
 
@@ -31,9 +11,10 @@ use anchor_spl::{
 };
 use vipers::*;
 
-#[error]
+#[error_code]
 pub enum ErrorCode {
     MyError,
+    MyOtherError,
 }
 
 #[account]
@@ -44,7 +25,7 @@ struct TestData {
 
 #[test]
 #[allow(deprecated)]
-pub fn test_compiles_deprecated() -> ProgramResult {
+pub fn test_compiles_deprecated() -> Result<()> {
     assert_keys!(token::ID, token::ID, "token program");
 
     Ok(())
@@ -52,7 +33,7 @@ pub fn test_compiles_deprecated() -> ProgramResult {
 
 #[test]
 #[allow(deprecated)]
-pub fn test_compiles() -> ProgramResult {
+pub fn test_compiles() -> Result<()> {
     let ata = get_associated_token_address(&token::ID, &token::ID);
     assert_ata!(ata, token::ID, token::ID, "ATA");
 
@@ -65,7 +46,7 @@ pub fn test_compiles() -> ProgramResult {
 
 #[test]
 #[allow(deprecated)]
-fn test_assert_owner() -> ProgramResult {
+fn test_assert_owner() -> Result<()> {
     let mut lamports: u64 = 8 + (TestData::default().try_to_vec().unwrap().len() as u64);
 
     let mut buffer: [u8; 16] = [0; 16];
@@ -88,7 +69,7 @@ fn test_assert_owner() -> ProgramResult {
 }
 
 #[test]
-fn test_unwrap_checked() -> ProgramResult {
+fn test_unwrap_checked() -> Result<()> {
     assert_throws!(
         {
             unwrap_checked!({
@@ -169,10 +150,9 @@ fn test_assert_keys_eq_pass() {
     });
 }
 
-use crate::anchor_lang::solana_program::program_pack::Pack;
+use anchor_lang::solana_program::program_pack::Pack;
 
 #[test]
-#[cfg(feature = "anchor-0_21_0")]
 fn test_assert_keys_eq_boxed() {
     let key = Pubkey::new_unique();
     let lamports = &mut 0;
@@ -246,5 +226,21 @@ fn test_assert_keys_neq_no_match() {
             )
         },
         ErrorCode::MyError
+    );
+}
+
+#[test]
+fn test_anchor_errors_eq_result() {
+    assert_eq!(
+        (err!(MyError) as Result<()>).into_cmp_error(),
+        (err!(MyError) as Result<()>).into_cmp_error(),
+    );
+}
+
+#[test]
+fn test_anchor_errors_ne_result() {
+    assert_ne!(
+        (err!(MyError) as Result<()>).into_cmp_error(),
+        (err!(MyOtherError) as Result<()>).into_cmp_error(),
     );
 }