+++ CHANGELOG.md
@@ -7,7 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Features
+
+- Adds `ComparableError`, which allows comparing errors.
+
+### Breaking
+
 - Make `spl-associated-token-account` an optional dependency ([#32](https://github.com/saber-hq/vipers/pull/32)).
+- Remove support for Anchor versions older than 0.22.
 
 ## [1.6.0] - 2022-02-14